Starting to flesh out the mock client more, from the bottom up.
Currently building through the client from what is needed to the working example.
The inmemory version needs to NOT use immutable components for that test to work
right. But I really need to pull that out into its own project as its fairly
useless. (Just use mockito right?) Anyways, the current build process is put in
the client what is needed, and then flesh out the functional test to get the whole
thing to work. At that point, I'll actually remove from the server-side impl the
in memory version since it is fairly useless.