Added the ability to include additional dependencies for the resolver.
This is needed in cases you want to resolve with utilities that need configured classes.
Example: DynamoDB needs to be injected really. You could do weird things with the
credentials but that's not the best way through this. You would not need this for
JDBC drivers as that can be done completely via string variables. But there are so
many ways to configure DDB that you may as well just get the real thing. Either way,
it adds to the capability.