Moved in my metrics wrapper lib to prepare for adding key service.
Before I can bring in the next set of code I have to add in the
metrics wrapper I built. The wrapper was designed so I can switch
to prometheus from drop-wizard metrics easily enough. Of course,
it turns out that wasn't so easy in the end so I'm going to have
to rebuild my prometheus client. But this atleast gets it working.