Finishing up the metrics refactoring... mostly.
I cleaned up the metrics and finished the helpers which is really
how they should be used. Updated the packages to be more inline
with a clean package tree. Still need to fully test it and give a
sample runtime to test the vendor implementations. Then I can
get the other vendor implementation done.