Fixes to the Cassandra testing process.
Turns out the @Rule for TestContainers rebuilds the containier each time.
This slows down the KeyDao to a crawl. Not enough to bring back the
cassandra support in database-test... but reconfigured the test so the
container is built once.

Using one keystore.cql file now. Added dagger components to get the test
working fine.