Cassandra DAO work: Preliminary
Finally got a plan for the Cassandra work. I reserve the right to
change how I'm dealing with 'active' keys but for now this is reasonable.
DynamoDB wants everything in one table, whereas Cassandra does not mind
the denormalization. I'd rather deal with deletions within Cassandra
than indexing a boolean. But I'll query other folks to verify that.

Note I still have not dealt with 'field level encryption/decryption'
for the keys itself. (To handle when database disks are removed and
the maintainer of the datastore doesn't deal with encryption correctly.)
That is coming next.

Finally, until the Cassandra DAO works to spec, the tests are broken
such that you cannot build the whole project in full. Hopefully I'll
have that working this weekend.