Refactored how binder works, and dagger.
Binder now builds within the binder manager, which means we can control it better than when it was all
via dagger.

Created a CQL Session module to manage the module now.

Finding the tests for cassandra slow on one computer, and fast on another. It is not a
hardware issue since the 'slow' computer has faster hardware. Need to deep dive this.
But effectively, there is a setup on each test. Its not the dao creation, nor container
restarts; neither are happening. Something else is going on.